third_party/WebKit/LayoutTests/fast/forms/fieldset/fieldset-type.html

<script src="../../../resources/js-test.js"></script>
<fieldset id="fs1" name="a"></fieldset>
<script>
description("This test checks fieldset.test attribute");
shouldBe('fs1.type', '"fieldset"');
</script>
